+++ packages/snaps-controllers/src/multichain/index.ts
@@ -1,1 +1,8 @@
-export * from './MultichainRouter';
+export { MultichainRouter } from './MultichainRouter';
+export type {
+  MultichainRouterGetSupportedAccountsAction,
+  MultichainRouterGetSupportedMethodsAction,
+  MultichainRouterHandleRequestAction,
+  MultichainRouterIsSupportedScopeAction,
+  MultichainRouterMethodActions,
+} from './MultichainRouter-method-action-types';